style: rename relationship to topology
